Remove full camera info from env spec; use float rewards and dictionary actions.

+++ src/robotamer/env_wrappers/rewards.py
@@ -5,5 +5,5 @@ class SparseRewardWrapper(gym.Wrapper):
 
     def step(self, action):
         obs, reward, done, info = self.env.step(action)
-        reward = int('success' in info and info['success'])
+        reward = float('success' in info and info['success'])
         return obs, reward, done, info
